Add airbyer property
