+++ airbyte-ci/connectors/pipelines/pipelines/commands/airbyte_ci.py
@@ -6,7 +6,6 @@
 
 import importlib
 from typing import List
-import sys
 
 import click
 from github import PullRequest
@@ -27,29 +26,8 @@ from .groups.connectors import connectors
 from .groups.metadata import metadata
 from .groups.tests import test
 
+from pipelines.telemetry import track_command
 
-import segment.analytics as analytics
-
-analytics.write_key = 'ER8EjdRVFut7n05XPaaTKrSEnjLscyKr'
-analytics.send = False
-def on_error(error, items):
-    print("An error occurred:", error)
-
-
-analytics.debug = True
-analytics.on_error = on_error
-
-def track_command(f):
-    def wrapper(*args, **kwargs):
-        full_cmd = " ".join(sys.argv)
-        is_local = kwargs.get('is_local', False)
-        user_id = 'local-user' if is_local else 'ci-user'
-
-        # IMPORTANT! do not log kwargs as they may contain secrets
-        analytics.track(user_id=user_id, event=f.__name__, properties={'command': full_cmd})
-
-        return f(*args, **kwargs)
-    return wrapper
 
 # HELPERS
